apps/ios: NotesModel.notes seam + demo sample buckets (Task 6)
